refactor: pgClient - pgService 으로 분리, 응답 형태 수정

+++ apps/commerce-api/src/main/java/com/loopers/application/payment/CallbackPaymentCommand.java
@@ -6,7 +6,7 @@ public record CallbackPaymentCommand(
         Long orderId,
         String cardType,
         String cardNo,
-        Long amount,
+        int amount,
         String status,
         String reason
 ) {
